Adding DiskManager and DiskIOScheduler (#474)
- Modifying StoreManager (now StorageManager) to instantiate a DiskManager for each disk on the node.
- Adding a DiskManager that manages all of the stores on a certain disk
- Adding a basic DiskIOScheduler class that is passed into each BlobStore instance. It is not yet used.
- Adding support for parallel startup of stores.
- Allowing store startup to still succeed even when individual stores fail to start up.

+++ ambry-clustermap/src/test/java/com.github.ambry.clustermap/MockReplicaId.java
@@ -103,8 +103,11 @@ public class MockReplicaId implements ReplicaId {
 
   public void cleanup() {
     File replicaDir = new File(replicaPath);
-    for (File replica : replicaDir.listFiles()) {
-      replica.delete();
+    File[] replicaDirFiles = replicaDir.listFiles();
+    if (replicaDirFiles != null) {
+      for (File replica : replicaDirFiles) {
+        replica.delete();
+      }
     }
     replicaDir.delete();
   }
